+++ config.inc.sample.php
@@ -43,6 +43,9 @@ $DatabaseName = 'database_name_here';
  * mysqldump for MySQL
  * @example /usr/bin/mysqldump
  * @example C:/xampp/mysql/bin/mysqldump.exe
+ *
+ * mariadb-dump for MariaDB
+ * @example /usr/bin/mariadb-dump
  */
 $DatabaseDumpPath = '';