+++ CHANGES.md
@@ -35,6 +35,7 @@ Changes in 11.6
 - Update French & Spanish translations in rosariosis.po
 - CSS responsive tables: Prevent overscroll/bounce in iOS 16+ MobileSafari and Chrome in zresponsive.css
 - CSS select: Normalize color on iOS 15+ in colors.css
+- MariaDB $DatabaseDumpPath allow 'mariadb-dump' binary in diagnostic.php & config.inc.sample.php
 
 Changes in 11.5.2
 -----------------